refactor: eslint on all contexts and hooks
