+++ src/contexts/MoviesContext.tsx
@@ -1,8 +1,13 @@
 import { createContext } from 'react'
 
-import { type MoviesContextType } from '../types'
+export interface MoviesContextType {
+  getMovie: (id: string) => Promise<any>
+  getCast: (id: string) => Promise<any>
+  getTrailer: (id: string) => Promise<any>
+  getProviders: (id: string) => Promise<any>
+}
 
-const MoviesContext = createContext<MoviesContextType>({} as MoviesContextType)
+const MoviesContext = createContext<MoviesContextType | null>(null)
 MoviesContext.displayName = 'MoviesContext'
 
 export default MoviesContext
