+++ src/contexts/ToastContext.tsx
@@ -1,8 +1,10 @@
 import { createContext } from 'react'
 
-import { type ToastContextType } from '../types'
+export interface ToastContextType {
+  showToast: (message: string, description: string, type: 'success' | 'error') => void
+}
 
-const ToastContext = createContext<ToastContextType>({} as ToastContextType)
+const ToastContext = createContext<ToastContextType | null>(null)
 ToastContext.displayName = 'ToastContext'
 
 export default ToastContext
