+++ src/hooks/index.ts
@@ -1,9 +1,7 @@
-import useAuth from './useAuth'
-import useData from './useData'
-import useMovies from './useMovies'
-import usePersonalData from './usePersonalData'
-import useTheme from './useTheme'
-import useToast from './useToast'
-import useUser from './useUser'
-
-export { useAuth, useData, useMovies, usePersonalData, useTheme, useToast, useUser }
+export { default as useAuth } from './useAuth'
+export { default as useData } from './useData'
+export { default as useMovies } from './useMovies'
+export { default as usePersonalData } from './usePersonalData'
+export { default as useTheme } from './useTheme'
+export { default as useToast } from './useToast'
+export { default as useUser } from './useUser'
