+++ src/hooks/useAuth.tsx
@@ -6,7 +6,7 @@ const useAuth = (): AuthContextType => {
   const userAuthContext = useContext(AuthContext)
 
   if (userAuthContext == null) {
-    throw new Error('useCurrentUser has to be used within <CurrentUserContext.Provider>')
+    throw new Error('useAuth has to be used within <AuthContext.Provider>')
   }
   return userAuthContext
 }
