+++ src/hooks/useData.tsx
@@ -1,9 +1,14 @@
 import { useContext } from 'react'
 
-import { DataContext } from '../contexts'
+import { DataContext, type DataContextType } from '../contexts'
 
-const useData = () => {
-  return useContext(DataContext)
+const useData = (): DataContextType => {
+  const useDataContext = useContext(DataContext)
+
+  if (useDataContext == null) {
+    throw new Error('useData has to be used within <DataContext.Provider>')
+  }
+  return useDataContext
 }
 
 export default useData
