+++ src/hooks/useMovies.tsx
@@ -1,9 +1,14 @@
 import { useContext } from 'react'
 
-import { MoviesContext } from '../contexts'
+import { MoviesContext, type MoviesContextType } from '../contexts'
 
-const useMovies = () => {
-  return useContext(MoviesContext)
+const useMovies = (): MoviesContextType => {
+  const useMoviesContext = useContext(MoviesContext)
+
+  if (useMoviesContext == null) {
+    throw new Error('useMovies has to be used within <MoviesContext.Provider>')
+  }
+  return useMoviesContext
 }
 
 export default useMovies
