+++ src/hooks/useTheme.tsx
@@ -1,9 +1,14 @@
 import { useContext } from 'react'
 
-import { ThemeContext } from '../contexts'
+import { ThemeContext, type ThemeContextType } from '../contexts'
 
-const useTheme = () => {
-  return useContext(ThemeContext)
+const useTheme = (): ThemeContextType => {
+  const useThemeContext = useContext(ThemeContext)
+
+  if (useThemeContext == null) {
+    throw new Error('useTheme has to be used within <ThemeContext.Provider>')
+  }
+  return useThemeContext
 }
 
 export default useTheme
