+++ src/hooks/useToast.tsx
@@ -1,9 +1,14 @@
 import { useContext } from 'react'
 
-import { ToastContext } from '../contexts'
+import { ToastContext, type ToastContextType } from '../contexts'
 
-const useToast = () => {
-  return useContext(ToastContext)
+const useToast = (): ToastContextType => {
+  const useToastContext = useContext(ToastContext)
+
+  if (useToastContext == null) {
+    throw new Error('useToast has to be used within <ToastContext.Provider>')
+  }
+  return useToastContext
 }
 
 export default useToast
